Started on delete button
I started working on the delete button, added it both in the UI and a funktion responsible for that button. I commented it out for now because it doesn't properly work. While it can delete the element in the UI, it doesn't get deleted in the list, so when you reload the list, the deleted line still is there. Need to find a way to delete in both the 4 Lists and the data.txt